[ICU-3346] proper cast to eliminate warning X-Commit-URL: https://ssl.icu-project.org/trac/changeset/13714 Former-commit-id: fc7071b708d0bd5c4e94253b3a772c523a458295

+++ icu4c/source/common/ucnvmbcs.c
@@ -970,7 +970,7 @@ _MBCSLoad(UConverterSharedData *sharedData,
                     MBCS_ENTRY_FINAL_ACTION(entry)==MBCS_STATE_CHANGE_ONLY &&
                     MBCS_ENTRY_FINAL_STATE(entry)!=0
                 ) {
-                    mbcsTable->dbcsOnlyState=MBCS_ENTRY_FINAL_STATE(entry);
+                    mbcsTable->dbcsOnlyState=(uint8_t)MBCS_ENTRY_FINAL_STATE(entry);
 
                     mbcsTable->outputType=MBCS_OUTPUT_DBCS_ONLY;
                 }
